+++ brainiak/matnormal/base.py
