+++ brainiak/matnormal/noise_cov_gp.py
