+++ brainiak/matnormal/noise_cov_kron.py
