+++ brainiak/matnormal/noise_covs.py
