+++ brainiak/matnormal/regression.py
