+++ tests/matnormal/matnorm_logp_conditional_test.py
@@ -2,9 +2,9 @@ import numpy as np
 from numpy.testing import assert_allclose
 from scipy.stats import wishart, multivariate_normal
 import tensorflow as tf
-from brainiak.matnorm_models.helpers import rmn
-from brainiak.matnorm_models.base import MatnormModelBase
-from brainiak.matnorm_models.noise_covs import *
+from brainiak.matnormal.helpers import rmn
+from brainiak.matnormal.base import MatnormModelBase
+from brainiak.matnormal.noise_covs import *
 
 # X is m x n, so A sould be m x p
